+++ packages/base/src/boot.js
@@ -1,7 +1,7 @@
 import whenDOMReady from "./util/whenDOMReady.js";
 import insertFontFace from "./FontFace.js";
 import { getTheme } from "./config/Theme.js";
-import { _applyTheme } from "./Theming.js";
+import applyTheme from "./theming/applyTheme.js";
 import whenPolyfillLoaded from "./compatibility/whenPolyfillLoaded.js";
 import { getFeature } from "./FeaturesRegistry.js";
 
@@ -19,7 +19,7 @@ const boot = () => {
 		}
 
 		await whenDOMReady();
-		await _applyTheme(getTheme());
+		await applyTheme(getTheme());
 		OpenUI5Support && OpenUI5Support.attachListeners();
 		insertFontFace();
 		await whenPolyfillLoaded();
