+++ packages/base/src/config/Theme.js
@@ -1,5 +1,5 @@
 import { getTheme as getConfiguredTheme } from "../InitialConfiguration.js";
-import { _applyTheme } from "../Theming.js";
+import applyTheme from "../theming/applyTheme.js";
 
 let theme;
 
@@ -19,7 +19,7 @@ const setTheme = async newTheme => {
 	theme = newTheme;
 
 	// Update CSS Custom Properties
-	await _applyTheme(theme);
+	await applyTheme(theme);
 };
 
 export {
