+++ packages/base/src/theming/ExternalThemePresent.js
@@ -1,0 +1,12 @@
+let externalThemePresent = false;
+
+const setExternalThemePresent = value => {
+	externalThemePresent = value;
+};
+
+const getExternalThemePresent = () => externalThemePresent;
+
+export {
+	getExternalThemePresent,
+	setExternalThemePresent,
+};
