+++ packages/base/src/theming/getEffectiveStyle.js
@@ -1,0 +1,15 @@
+import { getCustomCSS } from "./CustomStyle.js";
+
+const getEffectiveStyle = ElementClass => {
+	const tag = ElementClass.getMetadata().getTag();
+	const customStyle = getCustomCSS(tag) || "";
+	let componentStyles = ElementClass.styles;
+
+	if (Array.isArray(componentStyles)) {
+		componentStyles = componentStyles.join(" ");
+	}
+	return `${componentStyles} ${customStyle}`;
+};
+
+
+export default getEffectiveStyle;
