+++ packages/base/src/theming/injectThemeProperties.js
@@ -1,0 +1,27 @@
+import createStyleInHead from "../util/createStyleInHead.js";
+import { ponyfillNeeded, runPonyfill } from "./CSSVarsPonyfill.js";
+
+/**
+ * Creates/updates a style element holding all CSS Custom Properties
+ * @param cssText
+ * @param packageName
+ */
+const injectThemeProperties = (cssText, packageName) => {
+	// Needed for all browsers
+	const styleElement = document.head.querySelector(`style[data-ui5-theme-properties="${packageName}"]`);
+	if (styleElement) {
+		styleElement.textContent = cssText || "";	// in case of undefined
+	} else {
+		const attributes = {
+			"data-ui5-theme-properties": packageName,
+		};
+		createStyleInHead(cssText, attributes);
+	}
+
+	// When changing the theme, run the ponyfill immediately
+	if (ponyfillNeeded()) {
+		runPonyfill();
+	}
+};
+
+export default injectThemeProperties;
